BJhands-v2
Now the program detects a KeyDown instead of a Click on a button when entring the hand values.
Still possible to erase if you've made a mistake (before clicking on win or loss).
I also used a local int instead of an array to store the card values when entered so I removed the proba.cs file which was now useless.
The interface is really poor but whatever, for now I'm just focusing on the coding part.